+++ mirrord/agent/src/steal/http/filter.rs
@@ -1,21 +1,17 @@
-use std::{net::SocketAddr, sync::Arc};
+use std::{net::SocketAddr, sync::Arc, time::Duration};
 
 use dashmap::DashMap;
 use fancy_regex::Regex;
 use hyper::server::conn::http1;
 use mirrord_protocol::ConnectionId;
 use tokio::{io::copy_bidirectional, net::TcpStream, sync::mpsc::Sender};
-use tracing::error;
+use tracing::{error, trace};
 
-use super::{
-    error::HttpTrafficError, hyper_handler::HyperHandler, DefaultReversibleStream, HttpVersion,
-};
-use crate::{
-    steal::{http::error, HandlerHttpRequest},
-    util::ClientId,
-};
+use super::{hyper_handler::HyperHandler, DefaultReversibleStream, HttpVersion};
+use crate::{steal::HandlerHttpRequest, util::ClientId};
 
 const H2_PREFACE: &[u8] = b"PRI * HTTP/2.0";
+const DEFAULT_HTTP_VERSION_DETECTION_TIMEOUT: Duration = Duration::from_secs(10);
 
 /// Controls the amount of data we read when trying to detect if the stream's first message contains
 /// an HTTP request.
@@ -24,98 +20,41 @@ const H2_PREFACE: &[u8] = b"PRI * HTTP/2.0";
 /// requests.
 pub(super) const MINIMAL_HEADER_SIZE: usize = 10;
 
-/// Used to set up the creation of a [`HyperHandler`] task for the HTTP traffic stealer.
-#[derive(Debug)]
-pub(super) struct HttpFilterBuilder {
-    http_version: HttpVersion,
-    reversible_stream: DefaultReversibleStream,
+/// Read the start of the TCP stream, decide if it's HTTP (of a supported version), if it is, serve
+/// the connection with a [`HyperHandler`]. If it isn't, just forward the whole TCP connection to
+/// the original destination.
+#[tracing::instrument(
+    level = "trace",
+    skip(stolen_stream, matched_tx, connection_close_sender)
+)]
+pub(super) async fn filter_task(
+    stolen_stream: TcpStream,
     original_destination: SocketAddr,
     connection_id: ConnectionId,
-    client_filters: Arc<DashMap<ClientId, Regex>>,
+    filters: Arc<DashMap<ClientId, Regex>>,
     matched_tx: Sender<HandlerHttpRequest>,
-
-    /// For informing the stealer task that the connection was closed.
     connection_close_sender: Sender<ConnectionId>,
-}
-
-impl HttpFilterBuilder {
-    /// Does not consume bytes from the stream.
-    ///
-    /// Checks if the first available bytes in a stream could be of an http request.
-    ///
-    /// This is a best effort classification, not a guarantee that the stream is HTTP.
-    #[tracing::instrument(
-        level = "trace",
-        skip_all
-        fields(
-            local = ?stolen_stream.local_addr(),
-            peer = ?stolen_stream.peer_addr()
-        ),
-    )]
-    pub(super) async fn new(
-        stolen_stream: TcpStream,
-        original_destination: SocketAddr,
-        connection_id: ConnectionId,
-        filters: Arc<DashMap<ClientId, Regex>>,
-        matched_tx: Sender<HandlerHttpRequest>,
-        connection_close_sender: Sender<ConnectionId>,
-    ) -> Result<Self, HttpTrafficError> {
-        DefaultReversibleStream::read_header(stolen_stream)
-            .await
-            .map(|mut reversible_stream| {
-                let http_version = HttpVersion::new(
-                    reversible_stream.get_header(),
-                    &H2_PREFACE[..MINIMAL_HEADER_SIZE],
-                );
-
-                Self {
-                    http_version,
-                    client_filters: filters,
-                    reversible_stream,
-                    original_destination,
-                    connection_id,
-                    matched_tx,
-                    connection_close_sender,
-                }
-            })
-            .inspect_err(|fail| error!("Something went wrong in http filter {fail:#?}"))
-    }
-
-    /// Creates the hyper task, and returns an [`HttpFilter`] that contains the channels we use to
-    /// pass the requests to the layer.
-    #[tracing::instrument(
-        level = "debug",
-        skip(self),
-        fields(
-            self.http_version = ?self.http_version,
-            self.client_filters = ?self.client_filters,
-            self.connection_id = ?self.connection_id,
-            self.original_destination = ?self.original_destination,
-        )
-    )]
-    pub(super) fn start(self) -> Result<(), HttpTrafficError> {
-        let Self {
-            http_version,
-            client_filters,
-            matched_tx,
-            mut reversible_stream,
-            connection_id,
-            original_destination,
-            connection_close_sender,
-        } = self;
-
-        let port = original_destination.port();
-
-        match http_version {
-            HttpVersion::V1 => {
-                tokio::task::spawn(async move {
+) {
+    let port = original_destination.port();
+    match DefaultReversibleStream::read_header(
+        stolen_stream,
+        DEFAULT_HTTP_VERSION_DETECTION_TIMEOUT,
+    )
+    .await
+    {
+        Ok(mut reversible_stream) => {
+            match HttpVersion::new(
+                reversible_stream.get_header(),
+                &H2_PREFACE[..MINIMAL_HEADER_SIZE],
+            ) {
+                HttpVersion::V1 => {
                     // TODO: do we need to do something with this result?
                     let _res = http1::Builder::new()
                         .preserve_header_case(true)
                         .serve_connection(
                             reversible_stream,
                             HyperHandler {
-                                filters: client_filters,
+                                filters,
                                 matched_tx,
                                 connection_id,
                                 port,
@@ -133,24 +72,52 @@ impl HttpFilterBuilder {
                             error!("Main TcpConnectionStealer dropped connection close channel while HTTP filter is still running. \
                             Cannot report the closing of connection {connection_id}.");
                         });
-                });
-                Ok(())
-            }
-            // TODO(alex): hyper handling of HTTP/2 requires a bit more work, as it takes an
-            // "executor" (just `tokio::spawn` in the `Builder::new` function is good enough), and
-            // some more effort to chase some missing implementations.
-            HttpVersion::V2 | HttpVersion::NotHttp => {
-                let _passhtrough_task = tokio::task::spawn(async move {
-                    let mut interceptor_to_original =
-                        TcpStream::connect(original_destination).await?;
+                }
 
-                    copy_bidirectional(&mut reversible_stream, &mut interceptor_to_original)
-                        .await?;
-                    Ok::<_, error::HttpTrafficError>(())
-                });
-
-                Ok(())
+                // TODO(alex): hyper handling of HTTP/2 requires a bit more work, as it takes an
+                // "executor" (just `tokio::spawn` in the `Builder::new` function is good enough),
+                // and some more effort to chase some missing implementations.
+                HttpVersion::V2 | HttpVersion::NotHttp => {
+                    trace!(
+                        "Got a connection with unsupported protocol version, passing it through \
+                        to its original destination."
+                    );
+                    match TcpStream::connect(original_destination).await {
+                        Ok(mut interceptor_to_original) => {
+                            match copy_bidirectional(
+                                &mut reversible_stream,
+                                &mut interceptor_to_original,
+                            )
+                            .await
+                            {
+                                Ok((incoming, outgoing)) => {
+                                    trace!(
+                                        "Forwarded {incoming} incoming bytes and {outgoing} \
+                                        outgoing bytes in passthrough connection"
+                                    )
+                                }
+                                Err(err) => {
+                                    error!(
+                                        "Encountered error while forwarding unsupported \
+                                    connection to its original destination: {err:?}"
+                                    )
+                                }
+                            };
+                        }
+                        Err(err) => {
+                            error!(
+                                "Could not connect to original destination {original_destination:?}\
+                                 . Received a connection with an unsupported protocol version to a \
+                                 filtered HTTP port, but cannot forward the connection because of \
+                                 the connection error: {err:?}");
+                        }
+                    }
+                }
             }
+        }
+
+        Err(read_error) => {
+            error!("Got error while trying to read first bytes of TCP stream: {read_error:?}");
         }
     }
 }